feat(oyatsu): run incremental tick on index without opening workshop
Lift useOyatsuWorkshopRuntime to index (load + startLoop on mount), provide
shared state to OyatsuWorkshop via inject, and mount choice modal globally so
3-choice can appear before the panel is opened. Closing workshop no longer
stops the game loop.

Made-with: Cursor
